Feature/martin add delivrery process

See merge request cmc373/pld_agile!50

+++ code/src/main/java/view/DashBoardController.java
@@ -18,15 +18,13 @@ import javafx.stage.FileChooser;
 import lombok.Getter;
 import model.data.*;
 import org.apache.commons.lang.Validate;
+import org.apache.commons.lang.math.NumberUtils;
 
 import java.io.File;
 import java.net.URL;
 import java.sql.Time;
 import java.text.DecimalFormat;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.ResourceBundle;
+import java.util.*;
 
 public class DashBoardController implements Initializable, MapComponentInitializedListener {
 
@@ -44,13 +42,15 @@ public class DashBoardController implements Initializable, MapComponentInitializ
     }
 
     public void deleteDp() {
-        this.mainApp.deleteDp(deliveryProcessLoaded);
+        if(showConfirmationAlert("Are you sur you want to delete this Delivery Process ?")){
+            this.mainApp.deleteDp(deliveryProcessLoaded);
+        }
     }
 
     public void setActionPoints(final Tour tour) {
+        actionPointTableView.getSelectionModel().clearSelection();
         actionPoints.remove(0, actionPoints.size());
         actionPoints.addAll(tour.getActionPoints());
-
     }
 
 
@@ -79,7 +79,6 @@ public class DashBoardController implements Initializable, MapComponentInitializ
     private ObservableList<ActionPoint> actionPoints = FXCollections.observableArrayList();
 
     // Manage New DeliveryProcess
-    private DeliveryProcess newDeliveryProcess = null;
     private ActionPoint newPickUpActionPoint = null;
     private ActionPoint newDeliveryActionPoint = null;
     // Markers of new DeliveryProcess
@@ -105,10 +104,13 @@ public class DashBoardController implements Initializable, MapComponentInitializ
     private Label labelDeliveryCoordonates;
 
     @FXML
-    private TextField inputDeliveryTime;
-
+    private TextField inputDeliveryTimeH;
     @FXML
-    private TextField inputPickUpTime;
+    private TextField inputDeliveryTimeM;
+    @FXML
+    private TextField inputPickUpTimeH;
+    @FXML
+    private TextField inputPickUpTimeM;
 
     @FXML
     private Label dpNumber;
@@ -174,7 +176,7 @@ public class DashBoardController implements Initializable, MapComponentInitializ
                 cellData -> new SimpleStringProperty(cellData.getValue().getPassageTime()));
 
         actionPointTableView.getSelectionModel().selectedItemProperty().addListener(
-                (observable, oldValue, newValue) -> mainApp.showDeliveryProcess(newValue, tourLoaded));
+                (observable, oldValue, newValue) -> handelTableSelection(newValue));
 
         mapView.addMapInializedListener(this);
         mapView.setKey("AIzaSyDJDcPFKsYMTHWJUxVzoP0W7ERsx3Bhdgc");
@@ -218,11 +220,13 @@ public class DashBoardController implements Initializable, MapComponentInitializ
     // Create / Update
 
     public void createFakeActionPointList() {
-        List<ActionPoint> listActionPoints = new ArrayList<ActionPoint>();
+        List<ActionPoint> listActionPoints = new ArrayList<>();
         // Create a base actionPoint.
         ActionPoint base = new ActionPoint(tourLoaded.getStartTime(), tourLoaded.getBase(), ActionType.BASE);
+        ActionPoint end = new ActionPoint(tourLoaded.getStartTime(), tourLoaded.getBase(), ActionType.END);
 
         listActionPoints.add(base);
+        listActionPoints.add(end);
         for (DeliveryProcess deliveryProcess : tourLoaded.getDeliveryProcesses()) {
             listActionPoints.add(new ActionPoint(deliveryProcess.getPickUP()
                     .getTime(), deliveryProcess.getPickUP().getLocation(),
@@ -242,15 +246,29 @@ public class DashBoardController implements Initializable, MapComponentInitializ
     }
 
     public void addNewDeliveryProcess() {
-        if (canAdd()) {
-            //TODO Generate new DP
-            if (newDeliveryProcess != null) {
+        if (canAddDeliveryProcess()) {
+            if (newPickUpActionPoint != null && newDeliveryActionPoint != null) {
+                newPickUpActionPoint.setTime(parseStringToTime(inputPickUpTimeH.getText(), inputPickUpTimeM.getText()));
+                newDeliveryActionPoint.setTime(parseStringToTime(inputDeliveryTimeH.getText(), inputDeliveryTimeM.getText()));
+                this.mainApp.addDeliveryProcess(tourLoaded, newPickUpActionPoint, newDeliveryActionPoint);
+            } else {
+                showAlert("Action Imposible", "Error :", "The Delivery Process is not created", Alert.AlertType.ERROR);
             }
         } else {
             showAlert("Action Imposible", "Error :", "All the fields to create a delivery process are not completes", Alert.AlertType.ERROR);
         }
     }
 
+    private Time parseStringToTime(final String hours, final String minutes) {
+        Validate.notNull(hours, "hours null");
+        Validate.notNull(minutes, "minutes null");
+        Validate.isTrue(NumberUtils.isNumber(hours), "hours not a number");
+        Validate.isTrue(NumberUtils.isNumber(minutes), "minutes not a number");
+        Validate.isTrue(Integer.parseInt(hours) < 24 && Integer.parseInt(hours) >= 0, "not an hour");
+        Validate.isTrue(Integer.parseInt(minutes) < 59 && Integer.parseInt(minutes) >= 0, "not a minute");
+        final String toParse = hours + ":" + minutes + ":00";
+        return Time.valueOf(toParse);
+    }
     // Update view
 
     public void showDeliveryProcess(final DeliveryProcess deliveryProcess) {
@@ -264,23 +282,25 @@ public class DashBoardController implements Initializable, MapComponentInitializ
         dpDDuration.setText(deliveryDuration);
         dpPUDuration.setText(pickUpDuration);
         // actionPointTableView.setItems(null);
-        if (deliveryProcess.getPickUP().getActionType() == ActionType.BASE) {
-            dpDuration.setText(tourLoaded.getCompleteTime().toString());
-            List<Journey> journeyList = new ArrayList<Journey>();
-            journeyList.add(tourLoaded.getJourneyList().get(0));
-            displayMap();
-            drawAllActionPoints();
-            drawFullTour();
-            drawPolyline(getMCVPathFormJourneyListe(journeyList), "green", 0.5);
-            dPDistance.setText(String.valueOf(tourLoaded.getTotalDistance()));
-        } else {
-            // DISPLAY POLYLINE.
-            this.mainApp.getJourneyList(tourLoaded.getJourneyList(), deliveryProcess);
-            if (deliveryProcess.getTime() != null) {
-                dpDuration.setText(deliveryProcess.getTime().toString());
-            }
-            if (deliveryProcess.getDistance() != null) {
-                dPDistance.setText(String.valueOf(deliveryProcess.getDistance()));
+        // DISPLAY POLYLINE.
+        if (tourLoaded.getJourneyList() != null) {
+            if (deliveryProcess.getPickUP().getActionType() == ActionType.BASE) {
+                dpDuration.setText(tourLoaded.getCompleteTime().toString());
+                dPDistance.setText(String.valueOf(tourLoaded.getTotalDistance()) + " m");
+                List<Journey> journeyList = new ArrayList<Journey>();
+                journeyList.add(tourLoaded.getJourneyList().get(0));
+                displayMap();
+                drawAllActionPoints();
+                drawFullTour();
+                drawPolyline(getMCVPathFormJourneyListe(journeyList), "green", 0.5);
+            }else {
+                this.mainApp.getJourneyList(tourLoaded.getJourneyList(), deliveryProcess);
+                if (deliveryProcess.getTime() != null) {
+                    dpDuration.setText(deliveryProcess.getTime().toString());
+                }
+                if (deliveryProcess.getDistance() != null) {
+                    dPDistance.setText(String.valueOf(deliveryProcess.getDistance()) + " m");
+                }
             }
         }
     }
@@ -309,9 +329,10 @@ public class DashBoardController implements Initializable, MapComponentInitializ
 
         // Create a fake list of action Points To display.
         createFakeActionPointList();
-
+        actionPoints.remove(0, actionPoints.size());
         actionPoints.addAll(tourLoaded.getActionPoints());
 
+        System.out.println(actionPoints.size() + " action point size");
         // Add observable list data to the table
         actionPointTableView.setItems(actionPoints);
         drawAllActionPoints();
@@ -366,7 +387,7 @@ public class DashBoardController implements Initializable, MapComponentInitializ
             labelPickUpCoordonates.setText(stringFormater(actionPoint.getLocation()));
         }
         if (actionPoint.getActionType() == ActionType.DELIVERY) {
-            newPickUpActionPoint = actionPoint;
+            newDeliveryActionPoint = actionPoint;
             newDeliveryPointMarker = createMarker(actionPoint, MarkerType.DELIVERY);
             labelDeliveryCoordonates.setText(stringFormater(actionPoint.getLocation()));
         }
@@ -382,13 +403,14 @@ public class DashBoardController implements Initializable, MapComponentInitializ
     public void clearNewDeliveryProcess() {
         clearNewDeliveryPoint();
         clearNewPickUpPoint();
-        inputDeliveryTime.setText("");
-        inputPickUpTime.setText("");
+        inputDeliveryTimeM.setText("");
+        inputPickUpTimeM.setText("");
+        inputPickUpTimeH.setText("");
+        inputDeliveryTimeH.setText("");
         newPickUpPointMarker = null;
         newDeliveryPointMarker = null;
         newPickUpActionPoint = null;
         newDeliveryActionPoint = null;
-        newDeliveryProcess = null;
     }
 
     public void clearNewPickUpPoint() {
@@ -405,6 +427,7 @@ public class DashBoardController implements Initializable, MapComponentInitializ
 
     public void clearAll() {
         displayMap();
+        clearNewDeliveryProcess();
     }
     // Utils
 
@@ -458,11 +481,17 @@ public class DashBoardController implements Initializable, MapComponentInitializ
         return label.getText() == "";
     }
 
-    public Boolean canAdd() {
+    public Boolean canAddDeliveryProcess() {
+        if (inputPickUpTimeH.getText().equals("")) {
+            inputPickUpTimeH.setText("0");
+        }
+        if (inputDeliveryTimeH.getText().equals("")) {
+            inputDeliveryTimeH.setText("0");
+        }
         return labelDeliveryCoordonates.getText() != ""
                 && labelDeliveryCoordonates.getText() != ""
-                && inputPickUpTime.getText() != ""
-                && inputDeliveryTime.getText() != "";
+                && inputDeliveryTimeM.getText() != ""
+                && inputPickUpTimeM.getText() != "";
     }
 
     // Utils Pop Up
@@ -507,12 +536,45 @@ public class DashBoardController implements Initializable, MapComponentInitializ
         }
     }
 
+    private void handelTableSelection(ActionPoint newValue) {
+        if(newValue!= null && newValue.getActionType() == ActionType.END && tourLoaded.getJourneyList() != null){
+            dpDuration.setText(tourLoaded.getCompleteTime().toString());
+            dPDistance.setText(String.valueOf(tourLoaded.getTotalDistance()) + " m");
+            List<Journey> journeyList = new ArrayList<Journey>();
+            journeyList.add(tourLoaded.getJourneyList().get(tourLoaded.getJourneyList().size() - 1));
+            displayMap();
+            drawAllActionPoints();
+            drawFullTour();
+            drawPolyline(getMCVPathFormJourneyListe(journeyList), "black", 0.5);
+        }else {
+            mainApp.showDeliveryProcess(newValue, tourLoaded);
+        }
+    }
+
     private void showAlert(String title, String header, String msg, Alert.AlertType alertType) {
         Alert alert = new Alert(alertType);
         alert.setTitle(title);
         alert.setHeaderText(header);
         alert.setContentText(msg);
         alert.showAndWait();
+    }
+
+    private Boolean showConfirmationAlert(String msg) {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation");
+        alert.setHeaderText(msg);
+
+        ButtonType buttonTypeOne = new ButtonType("Yes");
+        ButtonType buttonTypeCancel = new ButtonType("Cancel", ButtonBar.ButtonData.CANCEL_CLOSE);
+
+        alert.getButtonTypes().setAll(buttonTypeOne, buttonTypeCancel);
+
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.get() == buttonTypeOne){
+            return true;
+        } else {
+            return false;
+        }
     }
 
     /**
